scripts/requirements.txt
# Google Generative AI SDK for Gemini, Imagen 3, and Veo 2
google-genai>=1.0.0
